test(onboard): split Model Router coverage
